Add AST node types:
This is not complete, just a good place to stop for now. There are some things
which are missing, such as functions, if statements, and so on. There is also
the issue of how to distinguish between expressions and statements, given that
an list of statements/expressions ending in an expression can be used as an
expression in a variable def. Still need to noodle out how to represent this
using AST nodes.